requirements.txt
python-telegram-bot==21.6
openai==1.51.0
mistralai==1.0.3
fpdf2==2.7.9
python-docx==1.1.2
pdfplumber==0.11.4
pandas==2.2.3
APScheduler==3.10.4
aiohttp==3.10.10
beautifulsoup4==4.12.3
python-dotenv==1.0.1
aiosqlite==0.20.0
Pillow==11.0.0
openpyxl==3.1.5
lxml==5.3.0
httpx==0.27.2
